Fix compilation errors in WiFi/MQTT managers and add session getter
- Fix min() type mismatch by casting constants to unsigned long
- Add get_current_session() public getter to GrindLogger
- Resolves type deduction errors in reconnection backoff logic

+++ src/network/mqtt_manager.cpp
@@ -406,7 +406,7 @@ void MQTTManager::handle_reconnect() {
 
     // Increment attempts and apply exponential backoff
     reconnect_attempts++;
-    reconnect_interval = min(reconnect_interval * 2, MQTT_MAX_RECONNECT_INTERVAL_MS);
+    reconnect_interval = min(reconnect_interval * 2, (unsigned long)MQTT_MAX_RECONNECT_INTERVAL_MS);
 
     Serial.printf("[MQTT] Reconnect attempt %d (next in %lums)\n",
                   reconnect_attempts, reconnect_interval);
